[refactor] Move the admission gates into their own module
The two questions the handshake asks before registering anyone — does any
member vouch for this peer, and do we agree on who founded the space — are
118 lines with a two-in, two-out seam: they need connectedPeers and the
logger, and the core needs isApprovedMember, resolveInvite and admitMember.
Sixteen imports moved out of swarm.js with them, which is the cohesion
argument. The event-taxonomy guard follows the divergence-clear call site
to its new file rather than pinning the old one.
